intentresolver: Add byte pagination to RequestBatcher and IntentResolver
Intent resolution batches are sequenced on raft and each batch can
consist of 100-200 intents. If an intent key or even value in some cases
are large, it is possible that resolving all intents in the batch would
result in a raft command size exceeding the max raft command size
kv.raft.command.max_size.

In PR #91976, we added support for TargetBytes for resolve intent and
resolve intent range raft commands.

In this PR, we set the TargetBytes field in the resolve intent and
resolve intent range requests in the RequestBatcher and IntentResolver
to half the max raft command size: kv.raft.command.max_size / 2. This
allows us to stop resolving intents in the batch as soon as we get close
to the max raft command size to lower the chance we exceed this limit.

Release note (ops change): Added support for a byte limit on resolve
intent and resolve intent range raft commands to prevent such commands
from exceeding the max raft command size.